Added alert controller for exceeding API request limit
