+++ YaMDS/Controllers/DetailViewController.swift
@@ -43,15 +43,5 @@ class DetailViewController: UIViewController {
     @IBAction func backButtonDidPressed(_ sender: UIButton) {
         dismiss(animated: true, completion: nil)
     }
-    
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
 
 }
